+++ experiments/training_ppo_sinusoid_discrete.py
@@ -8,8 +8,8 @@ for gpu in tf.config.experimental.list_physical_devices('GPU'):
 from keras import layers, models
 
 from finrock.data_feeder import PdDataFeeder
-from finrock.trading_env import TradingEnv
-from finrock.scalers import MinMaxScaler, ZScoreScaler
+from finrock.trading_env import TradingEnv, ActionSpace
+from finrock.scalers import MinMaxScaler, ZScoreScaler, LogReturnsScaler, Normalizer
 from finrock.reward import SimpleReward, AccountValueChangeReward
 from finrock.metrics import DifferentActions, AccountValue, MaxDrawdown, SharpeRatio
 from finrock.indicators import BolingerBands, RSI, PSAR, SMA, MACD
@@ -20,7 +20,7 @@ from rockrl.tensorflow import PPOAgent
 from rockrl.utils.vectorizedEnv import VectorizedEnv
 
 df = pd.read_csv('Datasets/random_sinusoid.csv')
-df = df[:-1000]
+df, df_test = df[:-1000], df[-1000:]
 
 
 pd_data_feeder = PdDataFeeder(
@@ -39,55 +39,104 @@ env = VectorizedEnv(
     env_object = TradingEnv,
     num_envs = num_envs,
     data_feeder = pd_data_feeder,
-    output_transformer = ZScoreScaler(),
+    output_transformer = Normalizer(),
     initial_balance = 1000.0,
     max_episode_steps = 1000,
-    window_size = 50,
+    window_size = 100,
     reward_function = AccountValueChangeReward(),
     metrics = [
         DifferentActions(),
         AccountValue(),
         MaxDrawdown(),
         SharpeRatio(),
-    ]
+    ],
+    # action_space = ActionSpace.SHORT_DISCRETE,
+    action_space = ActionSpace.DISCRETE,
 )
 
 action_space = env.action_space
 input_shape = env.observation_space.shape
 
+# def actor_model(input_shape, action_space):
+#     input = layers.Input(shape=input_shape, dtype=tf.float32)
+#     x = layers.Conv1D(filters=64, kernel_size=6, padding="same", activation="tanh")(input)
+#     x = layers.MaxPooling1D(pool_size=2)(x)
+#     x = layers.Conv1D(filters=32, kernel_size=3, padding="same", activation="tanh")(x)
+#     x = layers.MaxPooling1D(pool_size=2)(x)
+#     x = layers.Flatten()(x)
+#     # x = layers.Flatten()(input)
+#     # x = layers.Dense(512, activation='elu')(x)
+#     # x = layers.Dense(256, activation='elu')(x)
+#     x = layers.Dense(64, activation='elu')(x)
+#     # x = layers.Dropout(0.5)(x)
+#     output = layers.Dense(action_space, activation='softmax')(x) # discrete action space
+#     return models.Model(inputs=input, outputs=output)
+
+# def critic_model(input_shape):
+#     input = layers.Input(shape=input_shape, dtype=tf.float32)
+#     x = layers.Conv1D(filters=64, kernel_size=6, padding="same", activation="tanh")(input)
+#     x = layers.MaxPooling1D(pool_size=2)(x)
+#     x = layers.Conv1D(filters=32, kernel_size=3, padding="same", activation="tanh")(x)
+#     x = layers.MaxPooling1D(pool_size=2)(x)
+#     x = layers.Flatten()(x)
+#     # x = layers.Dense(512, activation='elu')(x)
+#     # x = layers.Dense(256, activation='elu')(x)
+#     x = layers.Dense(64, activation='elu')(x)
+#     # x = layers.Dropout(0.5)(x)
+#     output = layers.Dense(1, activation=None)(x)
+#     return models.Model(inputs=input, outputs=output)
+
+
+
 def actor_model(input_shape, action_space):
     input = layers.Input(shape=input_shape, dtype=tf.float32)
     x = layers.Flatten()(input)
-    x = layers.Dense(512, activation='elu')(x)
-    x = layers.Dense(256, activation='elu')(x)
-    x = layers.Dense(64, activation='elu')(x)
+    x = layers.Dense(512, activation='relu')(x)
+    x = layers.Dense(256, activation='relu')(x)
+    x = layers.Dense(64, activation='relu')(x)
     x = layers.Dropout(0.2)(x)
+    # x = layers.Bidirectional(layers.LSTM(128, return_sequences=True))(input)
+    # x = layers.Bidirectional(layers.LSTM(64, return_sequences=False))(x)
+    # x = layers.Flatten()(x)
+    # x = layers.Dense(64)(x)
     output = layers.Dense(action_space, activation='softmax')(x) # discrete action space
     return models.Model(inputs=input, outputs=output)
 
 def critic_model(input_shape):
     input = layers.Input(shape=input_shape, dtype=tf.float32)
     x = layers.Flatten()(input)
-    x = layers.Dense(512, activation='elu')(x)
-    x = layers.Dense(256, activation='elu')(x)
-    x = layers.Dense(64, activation='elu')(x)
+    x = layers.Dense(512, activation='relu')(x)
+    x = layers.Dense(256, activation='relu')(x)
+    x = layers.Dense(64, activation='relu')(x)
     x = layers.Dropout(0.2)(x)
+    # x = layers.Bidirectional(layers.LSTM(128, return_sequences=True))(input)
+    # x = layers.Bidirectional(layers.LSTM(64, return_sequences=False))(x)
+    # x = layers.Flatten()(x)
+    # x = layers.Dense(64)(x)
     output = layers.Dense(1, activation=None)(x)
     return models.Model(inputs=input, outputs=output)
 
 agent = PPOAgent(
     actor = actor_model(input_shape, action_space),
     critic = critic_model(input_shape),
-    optimizer=tf.keras.optimizers.Adam(learning_rate=0.0001),
-    batch_size=128,
-    lamda=0.95,
+    optimizer=tf.keras.optimizers.experimental.AdamW(learning_rate=0.0001),
+    batch_size=512,
+    lamda=0.90,
     kl_coeff=0.5,
     c2=0.01,
     writer_comment='ppo_sinusoid_discrete',
 )
+agent.summary()
 
 pd_data_feeder.save_config(agent.logdir)
 env.env.save_config(agent.logdir)
+
+# environment for testing
+test_env = TradingEnv.load_config(
+    PdDataFeeder.load_config(df_test, agent.logdir), 
+    agent.logdir
+)
+
 
 memory = MemoryManager(num_envs=num_envs)
 meanAverage = MeanAverage(best_mean_score_episode=1000)
@@ -108,7 +157,7 @@ while True:
         if meanAverage.is_best(agent.epoch):
             agent.save_models('ppo_sinusoid')
 
-        if history['kl_div'] > 0.05 and agent.epoch > 1000:
+        if history['kl_div'] > 0.10 and agent.epoch > 1000:
             agent.reduce_learning_rate(0.995, verbose=False)
 
         info = env_memory.infos[-1]
@@ -116,8 +165,20 @@ while True:
         agent.log_to_writer(info['metrics'])
         states[index], infos[index] = env.reset(index=index)
 
-    if agent.epoch >= 10000:
+        # test after last environment is done
+        if index == num_envs - 1:
+            test_state, test_info = test_env.reset()
+            while True:
+                action, _ = agent.act(test_state, training=False)
+                test_state, reward, terminated, truncated, info = test_env.step(action)
+
+                if terminated or truncated:
+                    for metric, value in info['metrics'].items():
+                        agent.log_to_writer({f'test/{metric}': value})
+                    break
+
+    if agent.epoch >= 20000:
         break
 
 env.close()
-exit()
+agent.close()