+++ finrock/scalers.py
@@ -74,3 +74,61 @@ class ZScoreScaler(Scaler):
         z_scores = np.nan_to_num((returns - np.mean(returns, axis=0)) / np.std(returns, axis=0))
 
         return z_scores
+    
+
+class LogReturnsScaler(Scaler):
+    def __init__(self):
+        super().__init__()
+
+    def logReturn(self, data):
+        logRet = np.log(data[1:] / data[:-1])
+        if np.isnan(logRet).any():
+            return np.nan_to_num(logRet) # nan to zero, but this is only workaround that may not be correct
+        return logRet
+
+    def transform(self, observations: Observations) -> np.ndarray:
+        scaled_data = [
+            self.logReturn(observations.open),
+            self.logReturn(observations.high),
+            self.logReturn(observations.low),
+            self.logReturn(observations.close),
+            self.logReturn(observations.volume),
+            observations.allocation_percentage[1:],
+        ]
+        indicators_data = []
+        for state in observations:
+            indicators_data.append([value for indicator in state.indicators for value in indicator["values"].values()])
+
+        for data in np.array(indicators_data).T:
+            scaled_data.append(self.logReturn(data))
+
+        return np.nan_to_num(np.array(scaled_data))
+    
+class Normalizer(Scaler):
+    """ Normalize data to be between 0 and 1
+    """
+    def __init__(self):
+        super().__init__()
+
+    def normalize(self, data):
+        """ Normalize batch data to be between 0 and 1
+        """
+        return (data - np.min(data)) / (np.max(data) - np.min(data))
+
+    def transform(self, observations: Observations) -> np.ndarray:
+        scaled_data = [
+            self.normalize(observations.open),
+            self.normalize(observations.high),
+            self.normalize(observations.low),
+            self.normalize(observations.close),
+            self.normalize(observations.volume),
+            observations.allocation_percentage,
+        ]
+        indicators_data = []
+        for state in observations:
+            indicators_data.append([value for indicator in state.indicators for value in indicator["values"].values()])
+
+        for data in np.array(indicators_data).T:
+            scaled_data.append(self.normalize(data))
+
+        return np.array(scaled_data)